Remove /summon and /users (I think 15+ years of being marked deprecated is enough, really)
git-svn-id: http://svn.inspircd.org/repository/trunk/inspircd@8088 e03df62e-2008-0410-955e-edbf42e46eb7

+++ include/ctables.h
@@ -72,10 +72,10 @@ class CoreExport command_t : public Extensible
 	int min_params;
 	/** used by /stats m
 	 */
-	long use_count;
+	long double use_count;
 	/** used by /stats m
  	 */
-	float total_bytes;
+	long double total_bytes;
 	/** used for resource tracking between modules
 	 */
 	std::string source;
